+++ src/bignum/bignum_rand_get.c
@@ -6,7 +6,7 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/22 16:59:14 by acazuc            #+#    #+#             */
-/*   Updated: 2018/08/10 20:06:52 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/10 21:27:25 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 uint64_t	g_bignum_rand;
 
-uint32_t	bignum_rand_get()
+uint64_t	bignum_rand_get()
 {
 	t_sha1_ctx	ctx;
 	uint8_t		digest[20];
@@ -26,5 +26,5 @@ uint32_t	bignum_rand_get()
 	sha1_init(&ctx);
 	sha1_update(&ctx, (uint8_t*)&g_bignum_rand, sizeof(g_bignum_rand));
 	sha1_final(digest, &ctx);
-	return (*(uint32_t*)digest);
+	return (*(uint64_t*)digest);
 }
